feat: dice analyzer
